working on litters only missing COMING SOON for empty litter
